release 1.19.0

+++ bot/helpers/handlers/handle_replicate_webhook.py
@@ -21,21 +21,22 @@ from bot.database.operations.request.updaters import update_request
 from bot.database.operations.transaction.writers import write_transaction
 from bot.database.operations.user.getters import get_user
 from bot.handlers.ai.face_swap_handler import PRICE_FACE_SWAP, handle_face_swap
-from bot.handlers.ai.flux_handler import PRICE_FLUX
+from bot.handlers.ai.flux_handler import PRICE_FLUX_1_DEV, PRICE_FLUX_1_PRO
 from bot.handlers.ai.music_gen_handler import PRICE_MUSIC_GEN, handle_music_gen
 from bot.handlers.ai.photoshop_ai_handler import (
+    PRICE_PHOTOSHOP_AI_UPSCALE,
     PRICE_PHOTOSHOP_AI_RESTORATION,
     PRICE_PHOTOSHOP_AI_COLORIZATION,
     PRICE_PHOTOSHOP_AI_REMOVAL_BACKGROUND,
     handle_photoshop_ai,
 )
-from bot.handlers.ai.stable_diffusion_handler import PRICE_STABLE_DIFFUSION
+from bot.handlers.ai.stable_diffusion_handler import PRICE_STABLE_DIFFUSION_XL, PRICE_STABLE_DIFFUSION_3
 from bot.helpers.senders.send_audio import send_audio
 from bot.helpers.senders.send_document import send_document
 from bot.helpers.senders.send_error_info import send_error_info
 from bot.helpers.senders.send_images import send_image
 from bot.helpers.updaters.update_user_usage_quota import update_user_usage_quota
-from bot.keyboards.common.common import build_reaction_keyboard, build_error_keyboard
+from bot.keyboards.common.common import build_reaction_keyboard, build_error_keyboard, build_buy_motivation_keyboard
 from bot.locales.main import get_user_language, get_localization
 from bot.locales.types import LanguageCode
 
@@ -100,10 +101,36 @@ async def handle_replicate_webhook(bot: Bot, dp: Dispatcher, prediction: dict):
             'seconds': generation.seconds,
         })
 
-    if product.details.get('quota') == Quota.STABLE_DIFFUSION:
-        asyncio.create_task(handle_replicate_stable_diffusion(bot, dp, user, user_language_code, request, generation))
-    elif product.details.get('quota') == Quota.FLUX:
-        asyncio.create_task(handle_replicate_flux(bot, dp, user, user_language_code, request, generation))
+    if (
+        product.details.get('quota') == Quota.STABLE_DIFFUSION_XL or
+        product.details.get('quota') == Quota.STABLE_DIFFUSION_3
+    ):
+        asyncio.create_task(
+            handle_replicate_stable_diffusion(
+                bot,
+                dp,
+                user,
+                user_language_code,
+                product.details.get('quota'),
+                request,
+                generation,
+            )
+        )
+    elif (
+        product.details.get('quota') == Quota.FLUX_1_DEV or
+        product.details.get('quota') == Quota.FLUX_1_PRO
+    ):
+        asyncio.create_task(
+            handle_replicate_flux(
+                bot,
+                dp,
+                user,
+                user_language_code,
+                product.details.get('quota'),
+                request,
+                generation,
+            )
+        )
     elif product.details.get('quota') == Quota.FACE_SWAP:
         asyncio.create_task(handle_replicate_face_swap(bot, dp, user, user_language_code, request, generation))
     elif product.details.get('quota') == Quota.PHOTOSHOP_AI:
@@ -144,7 +171,6 @@ async def handle_replicate_photoshop_ai(
             chat_id=user.telegram_chat_id,
             text=get_localization(user_language_code).ERROR,
             reply_markup=reply_markup,
-            parse_mode=None,
         )
 
     if request.status != RequestStatus.FINISHED:
@@ -154,7 +180,9 @@ async def handle_replicate_photoshop_ai(
         })
 
         action_name = request.details.get('type')
-        if action_name == PhotoshopAIAction.RESTORATION:
+        if action_name == PhotoshopAIAction.UPSCALE:
+            total_price = round(PRICE_PHOTOSHOP_AI_UPSCALE * generation.seconds, 6)
+        elif action_name == PhotoshopAIAction.RESTORATION:
             total_price = round(PRICE_PHOTOSHOP_AI_RESTORATION * generation.seconds, 6)
         elif action_name == PhotoshopAIAction.COLORIZATION:
             total_price = round(PRICE_PHOTOSHOP_AI_COLORIZATION * generation.seconds, 6)
@@ -364,7 +392,6 @@ async def handle_replicate_music_gen(
             chat_id=user.telegram_chat_id,
             text=get_localization(user_language_code).ERROR,
             reply_markup=reply_markup,
-            parse_mode=None,
         )
 
     if request.status != RequestStatus.FINISHED:
@@ -424,15 +451,17 @@ async def handle_replicate_stable_diffusion(
     dp: Dispatcher,
     user: User,
     user_language_code: LanguageCode,
+    user_quota: Quota,
     request: Request,
     generation: Generation,
 ):
+    is_suggestion = generation.details.get('is_suggestion', False)
     prompt = generation.details.get('prompt')
 
-    if generation.result:
-        footer_text = f'\n\n🖼 {user.daily_limits[Quota.STABLE_DIFFUSION] + user.additional_usage_quota[Quota.STABLE_DIFFUSION]}' \
+    if generation.result and not is_suggestion:
+        footer_text = f'\n\n🖼 {user.daily_limits[user_quota] + user.additional_usage_quota[user_quota]}' \
             if user.settings[Model.STABLE_DIFFUSION][UserSettings.SHOW_USAGE_QUOTA] and \
-               user.daily_limits[Quota.STABLE_DIFFUSION] != float('inf') else ''
+               user.daily_limits[user_quota] != float('inf') else ''
         caption = f'{get_localization(user_language_code).GENERATION_IMAGE_SUCCESS}{footer_text}'
 
         reply_markup = build_reaction_keyboard(generation.id)
@@ -440,19 +469,29 @@ async def handle_replicate_stable_diffusion(
             await send_document(bot, user.telegram_chat_id, generation.result, reply_markup, caption)
         else:
             await send_image(bot, user.telegram_chat_id, generation.result, reply_markup, caption)
+    elif generation.result and is_suggestion:
+        header_text = f'{get_localization(user_language_code).example_image_model(get_localization(user_language_code).STABLE_DIFFUSION_3)}\n'
+        footer_text = f'\n{get_localization(user_language_code).EXAMPLE_INFO}'
+        full_text = f'{header_text}{footer_text}'
+        await send_image(
+            bot,
+            user.telegram_chat_id,
+            generation.result,
+            build_buy_motivation_keyboard(user_language_code),
+            full_text,
+        )
     elif generation.has_error:
-        await bot.send_sticker(
-            chat_id=user.telegram_chat_id,
-            sticker=config.MESSAGE_STICKERS.get(MessageSticker.ERROR),
-        )
+        if not is_suggestion:
+            await bot.send_sticker(
+                chat_id=user.telegram_chat_id,
+                sticker=config.MESSAGE_STICKERS.get(MessageSticker.ERROR),
+            )
 
-        reply_markup = build_error_keyboard(user_language_code)
-        await bot.send_message(
-            chat_id=user.telegram_chat_id,
-            text=get_localization(user_language_code).ERROR,
-            reply_markup=reply_markup,
-            parse_mode=None,
-        )
+            await bot.send_message(
+                chat_id=user.telegram_chat_id,
+                text=get_localization(user_language_code).ERROR,
+                reply_markup=build_error_keyboard(user_language_code),
+            )
 
     if request.status != RequestStatus.FINISHED:
         request.status = RequestStatus.FINISHED
@@ -460,7 +499,7 @@ async def handle_replicate_stable_diffusion(
             'status': request.status
         })
 
-        total_price = PRICE_STABLE_DIFFUSION
+        total_price = PRICE_STABLE_DIFFUSION_XL if user_quota == Quota.STABLE_DIFFUSION_XL else PRICE_STABLE_DIFFUSION_3
         update_tasks = [
             write_transaction(
                 user_id=user.id,
@@ -473,15 +512,20 @@ async def handle_replicate_stable_diffusion(
                 details={
                     'result': generation.result,
                     'prompt': prompt,
+                    'is_suggestion': is_suggestion,
                     'has_error': generation.has_error,
                 },
             ),
-            update_user_usage_quota(
-                user,
-                Quota.STABLE_DIFFUSION,
-                1 if generation.result else 0,
-            ),
         ]
+
+        if generation.result and not is_suggestion:
+            update_tasks.append(
+                update_user_usage_quota(
+                    user,
+                    user_quota,
+                    1,
+                ),
+            )
 
         await asyncio.gather(*update_tasks)
 
@@ -495,11 +539,12 @@ async def handle_replicate_stable_diffusion(
         )
         await state.clear()
 
-        for processing_message_id in request.processing_message_ids:
-            try:
-                await bot.delete_message(user.telegram_chat_id, processing_message_id)
-            except Exception:
-                continue
+        if not is_suggestion:
+            for processing_message_id in request.processing_message_ids:
+                try:
+                    await bot.delete_message(user.telegram_chat_id, processing_message_id)
+                except Exception:
+                    continue
 
 
 async def handle_replicate_flux(
@@ -507,15 +552,17 @@ async def handle_replicate_flux(
     dp: Dispatcher,
     user: User,
     user_language_code: LanguageCode,
+    user_quota: Quota,
     request: Request,
     generation: Generation,
 ):
+    is_suggestion = generation.details.get('is_suggestion', False)
     prompt = generation.details.get('prompt')
 
-    if generation.result:
-        footer_text = f'\n\n🖼 {user.daily_limits[Quota.FLUX] + user.additional_usage_quota[Quota.FLUX]}' \
+    if generation.result and not is_suggestion:
+        footer_text = f'\n\n🖼 {user.daily_limits[user_quota] + user.additional_usage_quota[user_quota]}' \
             if user.settings[Model.FLUX][UserSettings.SHOW_USAGE_QUOTA] and \
-               user.daily_limits[Quota.FLUX] != float('inf') else ''
+               user.daily_limits[user_quota] != float('inf') else ''
         caption = f'{get_localization(user_language_code).GENERATION_IMAGE_SUCCESS}{footer_text}'
 
         reply_markup = build_reaction_keyboard(generation.id)
@@ -523,19 +570,30 @@ async def handle_replicate_flux(
             await send_document(bot, user.telegram_chat_id, generation.result, reply_markup, caption)
         else:
             await send_image(bot, user.telegram_chat_id, generation.result, reply_markup, caption)
+    elif generation.result and is_suggestion:
+        header_text = f'{get_localization(user_language_code).example_image_model(get_localization(user_language_code).FLUX_1_PRO)}\n'
+        footer_text = f'\n{get_localization(user_language_code).EXAMPLE_INFO}'
+        full_text = f'{header_text}{footer_text}'
+        await send_image(
+            bot,
+            user.telegram_chat_id,
+            generation.result,
+            build_buy_motivation_keyboard(user_language_code),
+            full_text,
+        )
     elif generation.has_error:
-        await bot.send_sticker(
-            chat_id=user.telegram_chat_id,
-            sticker=config.MESSAGE_STICKERS.get(MessageSticker.ERROR),
-        )
+        if not is_suggestion:
+            await bot.send_sticker(
+                chat_id=user.telegram_chat_id,
+                sticker=config.MESSAGE_STICKERS.get(MessageSticker.ERROR),
+            )
 
-        reply_markup = build_error_keyboard(user_language_code)
-        await bot.send_message(
-            chat_id=user.telegram_chat_id,
-            text=get_localization(user_language_code).ERROR,
-            reply_markup=reply_markup,
-            parse_mode=None,
-        )
+            reply_markup = build_error_keyboard(user_language_code)
+            await bot.send_message(
+                chat_id=user.telegram_chat_id,
+                text=get_localization(user_language_code).ERROR,
+                reply_markup=reply_markup,
+            )
 
     if request.status != RequestStatus.FINISHED:
         request.status = RequestStatus.FINISHED
@@ -543,7 +601,7 @@ async def handle_replicate_flux(
             'status': request.status
         })
 
-        total_price = PRICE_FLUX
+        total_price = PRICE_FLUX_1_DEV if user_quota == Quota.FLUX_1_DEV else PRICE_FLUX_1_PRO
         update_tasks = [
             write_transaction(
                 user_id=user.id,
@@ -556,15 +614,20 @@ async def handle_replicate_flux(
                 details={
                     'result': generation.result,
                     'prompt': prompt,
+                    'is_suggestion': is_suggestion,
                     'has_error': generation.has_error,
                 },
             ),
-            update_user_usage_quota(
-                user,
-                Quota.FLUX,
-                1 if generation.result else 0,
-            ),
         ]
+
+        if generation.result and not is_suggestion:
+            update_tasks.append(
+                update_user_usage_quota(
+                    user,
+                    user_quota,
+                    1,
+                ),
+            )
 
         await asyncio.gather(*update_tasks)
 
@@ -578,8 +641,9 @@ async def handle_replicate_flux(
         )
         await state.clear()
 
-        for processing_message_id in request.processing_message_ids:
-            try:
-                await bot.delete_message(user.telegram_chat_id, processing_message_id)
-            except Exception:
-                continue
+        if not is_suggestion:
+            for processing_message_id in request.processing_message_ids:
+                try:
+                    await bot.delete_message(user.telegram_chat_id, processing_message_id)
+                except Exception:
+                    continue
